added Climber move up command
