+++ src/main/java/frc/robot/commands/climber/Elevator.java
@@ -1,0 +1,28 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package frc.robot.commands.climber;
+
+import edu.wpi.first.wpilibj2.command.InstantCommand;
+import frc.robot.subsystems.Climber;
+
+// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
+// information, see:
+// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
+public class Elevator extends InstantCommand {
+  private Climber climber;
+  
+  public Elevator(Climber climber) {
+    // Use addRequirements() here to declare subsystem dependencies.
+    this.climber = climber;
+    addRequirements(climber);
+  }
+
+  // Called when the command is initially scheduled.
+  @Override
+  public void initialize() {
+    climber.m_climbFront.set(-0.1);
+    climber.m_climbBack.set(0.1);
+  }
+}
