+++ src/main/java/frc/robot/commands/climber/ElevatorUp.java
@@ -2,23 +2,20 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems;
+package frc.robot.commands.climber;
 
-import edu.wpi.first.wpilibj2.command.InstantCommand;
+import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import frc.robot.subsystems.Climber;
+import frc.robot.commands.climber.Elevator;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class ElevatorUp extends InstantCommand {
-  private Climber climber;
-  
+public class ElevatorUp extends SequentialCommandGroup {
+  /** Creates a new ElevatorUp. */
   public ElevatorUp(Climber climber) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    this.climber = climber;
-    addRequirements(climber);
+    // Add your commands in the addCommands() call, e.g.
+    // addCommands(new FooCommand(), new BarCommand());
+    addCommands(new Elevator(climber));
   }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {}
 }
